StarDelegate: Fix: Transition from "mouse edit mode" to "keyboard edit mode"
Correctly handle the case of a user pressing the edit key while the
mouse is also hovering over the focused cell.

Without this special handling, the pressing the edit key on a cell
would simply be ignored if the mouse is hovering over that cell,
which would most certainly break user expectations.