StarDelegate (3/4): Restore "mouse edit mode" when exiting "keyboard+mouse edit mode"
This is important for the case when all of the following is true:

* The user is currently editing a star rating cell via the keyboard.
* The mouse cursor is inside that cell.
* The user finishes editing by pressing the Return key or the Escape key.

Without the special handling in place, the user would have to move
the mouse outside the cell and then back in to see the cell react
to the movements of the mouse cursor again.

                ====================
                = More information =
                ====================

The StarRating control can be edited via mouse, keyboard, or
both at the same time. The following transitions are currently
handled correctly (marked by ✓) resp. affected by this change
(marked by =>):

       [✓] "Not editing"       -> "Edit via mouse"          (on mouseEnter)
       [✓] "Not editing"       -> "Edit via keyboard"       (on editKeyPressed)

       [✓] "Edit via mouse"    -> "Not editing"             (on mouseLeave)
       [✓] "Edit via mouse"    -> "Edit via keyboard+mouse" (on editKeyPressed)

       [✓] "Edit via keyboard" -> "Not editing"             (on exitKeyPressed)
       [✓] "Edit via keyboard" -> "Edit via keyboard+mouse" (on mouseEnter)

[ ] => [✓] "Edit via keyboard+mouse" -> "Edit via mouse"    (on exitKeyPressed)
       [~] "Edit via keyboard+mouse" -> "Edit via keyboard" (on mouseLeave)