StarDelegate: Fix: Defer the restore action to avoid breaking QAbstractItemView
QAbstractItemView doesn't like it when we try to open an editor while
the closing of the previously open editor is not fully finished,
because that messes up its internal state.

We avoid this race condition by deferring the restore until we have
returned to the event loop (using basically the same technique as
QObject::deleteLater().