StarDelegate: Fix: Workaround for bug of closePersistentEditor
Contrary to its name, the QAbstractItemView::closePersistentEditor()
method will also close a non-persistent editor which is present at
the specified QModelIndex.

If we do not detect this case, moving the mouse outside of a star
rating editor that is currently in "keyboard edit mode" would both
involuntarily exit the edit mode and break our internal state.