Fix autoExpand timer reset behavior (workaround for bug in Qt framework)
Starting with at least Qt 5.0.0 (released in 2011) and still present
in current versions of Qt (Qt 6.8.0 at the time of this commit), there
is a bug in the implementation of QTreeView::dragMoveEvent and autoExpandDelay:

QT BUG DESCRIPTION

Instead of resetting the delay timer whenever the mouse moves to a
new item, it is reset on every little mouse movement, which makes
autoExpand useless e.g. on laptop touchpads.

OUR WORKAROUND

Only reset the delay timer whenever the mouse has moved to a new item,
by bypassing QTreeView::dragMoveEvent() and directly calling
QAbstractItemView::dragMoveEvent() instead unless the mouse
has moved to a new item.