Schema: Relax the name uniqueness constraint for crates and crate folders
Instead of requiring all names to be globally unique, we
require them to be unique within a given parent folder.

This requires updating the constraints of the "crates" table,
which in turn leads to recreating the table, due to SQLite's limited
implementation of the ALTER TABLE statement.

See the SQLite documentation for more guidance:
https://www.sqlite.org/lang_altertable.html#making_other_kinds_of_table_schema_changes

As a sidenote: It is allowed to have a playlist and a folder with
the same name inside a single parent folder, i.e. playlist and folder
names do not conflict.